Release v4.0.0

+++ composeApp/src/commonMain/kotlin/PreviewScreenUiAction.kt
@@ -5,7 +5,7 @@ sealed interface PreviewScreenUiAction {
 
     data class SetAspectRatio(val value: Float) : PreviewScreenUiAction
     data class SetCornerRadius(val value: Float) : PreviewScreenUiAction
-    data class SetCornerSmoothing(val value: Float) : PreviewScreenUiAction
+    data class SetCornerSmoothing(val value: Int) : PreviewScreenUiAction
 
     data object Reset : PreviewScreenUiAction
     data object Random : PreviewScreenUiAction
